Add files via upload
I am still unsure how to correctly prove secrecy in my first attempt to model the protocol:
At the moment, it seems that the variables I check secrecy on are some sort of globally declared, while the variables I out() in the process are only locally named. Therefore proverif thinks secrecy holds, while it should in fact not hold for gx and gy.